Add clang-format support
I've attempted to match the existing formatting as closely as possible, but
as some of the formatting is internally inconsistent, there are a number
of changes that unfortunately will make git blame a bit confusing. There are
a few options here I don't think improve things like AllowShortBlocksOnASingleLine
causing a few more complex statements to be collapsed onto a single line,
but overall this change should unify the code without too much confusion.

As a manual step I've also trimmed all trailing spaces in .h and .c files
as apparently clang-format only does this for lines it modifies.